warnings

+++ src/tutorial/laplace_double_integral.cpp
@@ -38,7 +38,7 @@ void tester(func_space const &w)
 	using namespace boost::math::double_constants;
 	
 	// compute number of DOF and allocate result matrix
-	int nDOF = w.get_num_dofs();
+	size_t nDOF = w.get_num_dofs();
 	dMatrix I(nDOF, nDOF);
 	I.setZero();
